Integrated imgui to the extent that build goes through successfully
Integration of imgui into this project and all the little steps that this ensue.
The integration and the work for this template is not all the way done, as there
is no actual example code or interaction with imgui from the code in 'src' folder yet.
All I've done so far is integrate it with my cmake buildsystem so it builds without errors.
A lot of help/guidance was taken from the guide I've added link to in the README.md file.
Heres a rough summary of the steps taken so far:

- Add imgui as a git-submodule (Alternative to adding the source files individually
  to the project. This way it's smoother to upgrade to a newer version of imgui for the future)

- Add SDL2 and OpenGL as dependencies to this project (With imgui the integrator has to choose
  what to use for the backend. In this case I went for SDL2 for the event handling and OpenGL
  for the rendering. See more info in the imgui repository)

- To add OpenGL we also need GLAD (Multi-Language GL/GLES/EGL/GLX/WGL Loader-Generator) is an
  OpenGL loading library that loads pointers to OpenGL functions at runtime, and so this is
  required so the OpenGL functions can be accessed.
  This is as suppose to imgui added directly as individual files instead of as a git-submodule.